rename to sequenceS file name
